requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
openai>=0.27,<1.0
openai-whisper>=1.1,<2.0
pyannote-audio>=2.1,<3.0
python-dotenv>=1.0,<2.0
torch>=1.13,<2.0
torchaudio>=0.13,<2.0
transformers>=4.30,<5.0